news_collection/requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
Pillow>=10.0.0
